#' Reopens a TileDB array an opened tiledb array #' #' Reopening an array is useful when the array got updated after it got opened #' and the tiledb array object got created. To sync-up with the updates, #' the user must either close the array and open again, #' or just use tiledb_reopen(array) which can be faster because #' only metdata regarding updates has to be loaded. #' #' @param object tileb array object #' @return the reopened array object #' @export setGeneric("reopen", function(object, ...) standardGeneric("reopen"))

#' @export setMethod("reopen", "tiledb_dense", function(object) libtiledb_array_reopen(object@ptr) return(object) )

# S4 method for tiledb_dense
is.sparse(object)

Arguments

object
tiledb_dense

...
Extra parameter for method signature, currently unused.

Value

FALSE

Details

#' Closes a tiledb array object #' #' @param conn tiledb array object #' @return returns the closed array object close.tiledb_dense <- function(conn, ...) stopifnot(is(conn, "tiledb_dense")) libtiledb_array_close(conn@ptr) return(conn); Returns true is if the array or array_schema is sparse