Constructs a tiledb_array object backed by a persisted tiledb array uri

tiledb_array returns a new object. This class is experimental.

Usage

tiledb_array(
  uri,
  query_type = c("READ", "WRITE"),
  is.sparse = NA,
  as.data.frame = FALSE,
  attrs = character(),
  extended = TRUE,
  selected_ranges = list(),
  query_layout = character(),
  datetimes_as_int64 = FALSE,
  encryption_key = character(),
  timestamp = as.POSIXct(double(), origin = "1970-01-01"),
  as.matrix = FALSE,
  as.array = FALSE,
  query_condition = new("tiledb_query_condition"),
  timestamp_start = as.POSIXct(double(), origin = "1970-01-01"),
  timestamp_end = as.POSIXct(double(), origin = "1970-01-01"),
  return_as = get_return_as_preference(),
  query_statistics = FALSE,
  strings_as_factors = getOption("stringsAsFactors", FALSE),
  sil = list(),
  dumpbuffers = character(),
  buffers = list(),
  ctx = tiledb_get_context()
)

tiledb_dense(...)

tiledb_sparse(...)

Arguments

uri

uri path to the tiledb dense array

query_type

optionally loads the array in "READ" or "WRITE" only modes.

is.sparse

optional logical switch, defaults to "NA" letting array determine it

as.data.frame

optional logical switch, defaults to "FALSE"

attrs

optional character vector to select attributes, default is empty implying all are selected, the special value NA_character_ has the opposite effect and implies no attributes are returned.

extended

optional logical switch selecting wide ‘data.frame’ format, defaults to TRUE

selected_ranges

optional A list with matrices where each matrix i describes the (min,max) pair of ranges for dimension i

query_layout

optional A value for the TileDB query layout, defaults to an empty character variable indicating no special layout is set

datetimes_as_int64

optional A logical value selecting date and datetime value representation as ‘raw’ integer64 and not as Date, POSIXct or nanotime objects.

encryption_key

optional A character value with an AES-256 encryption key in case the array was written with encryption.

timestamp

optional A POSIXct Datetime value determining where in time the array is to be openened. Deprecated, use ‘timestamp_start’ instead

as.matrix

optional logical switch, defaults to "FALSE"; currently limited to dense matrices; in the case of multiple attributes in query a list of matrices is returned

as.array

optional logical switch, defaults to "FALSE"; in the case of multiple attributes in query a list of arrays is returned

query_condition

optional tiledb_query_condition object, by default uninitialized without a condition; this functionality requires TileDB 2.3.0 or later

timestamp_start

optional A POSIXct Datetime value determining the inclusive time point at which the array is to be openened. No fragments written earlier will be considered.

timestamp_end

optional A POSIXct Datetime value determining the inclusive time point until which the array is to be openened. No fragments written earlier later be considered.

return_as

optional A character value with the desired tiledb_array conversion, permitted values are ‘asis’ (default, returning a list of columns), ‘array’, ‘matrix’,‘data.frame’, ‘data.table’ or ‘tibble’; the latter two require the respective packages installed. The existing as.* arguments take precedent over this.

query_statistics

optional A logical value, defaults to ‘FALSE’; if ‘TRUE’ the query statistics are returned (as a JSON string) via the attribute ‘query_statistics’ of the return object.

strings_as_factors

An optional logical to convert character columns to factor type; defaults to the value of getOption("stringsAsFactors", FALSE).

sil

optional A list, by default empty to store schema information when query objects are parsed.

dumpbuffers

An optional character variable with a directory name (relative to /dev/shm) for writing out results buffers (for internal use / testing)

buffers

An optional list with full pathnames of shared memory buffers to read data from

ctx

optional tiledb_ctx

...

Used as a pass-through for tiledb_dense and tiledb_sparse aliasing

Value

tiledb_array object