Constructs a tiledb_sparse object backed by a persisted tiledb array uri

tiledb_sparse returns a list of coordinates and attributes vectors for reads

tiledb_sparse(
  uri,
  query_type = c("READ", "WRITE"),
  as.data.frame = FALSE,
  attrs = character(),
  extended = TRUE,
  ctx = tiledb_get_context()
)

Arguments

uri
uri path to the tiledb dense array

query_type
optionally loads the array in "READ" or "WRITE" only modes.

as.data.frame
optional logical switch, defaults to "FALSE"

attrs
optional character vector to select attributes, default is empty implying all are selected

extended
optional logical switch selecting wide ‘data.frame’ format, defaults to "TRUE"

ctx
tiledb_ctx (optional)

Value

tiledb_sparse array object

Planned Deprecation

We plan to deprecate the tiledb_sparse array type in a future release. While exact timelines have not been finalised, it is advised to the tiledb_array for both dense and sparse arrays going forward.